requirements.txt
click==6.7
factory-boy==2.8.1
Faker==0.7.11
Flask==0.12.1
Flask-Cors==3.0.2
Flask-JSON==0.3.0
Flask-JWT==0.3.2
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
passlib==1.7.1
psycopg2==2.7.1
py==1.4.33
PyJWT==1.4.2
pytest==3.0.7
python-dateutil==2.6.0
six==1.10.0
SQLAlchemy==1.1.9
voluptuous==0.10.5
Werkzeug==0.12.1
